+++ Cargo.toml
@@ -15,13 +15,13 @@ harness = false
 path = "benches/bench.rs"
 
 [features]
-default = []
+default = ["rustc-serialize"]
 nightly = []
-iron-handlers = ["iron"]
+iron-handlers = ["iron", "rustc-serialize"]
 expose-test-schema = []
 
 [dependencies]
-rustc-serialize = "^0.3.19"
+rustc-serialize = { version = "^0.3.19", optional = true }
 iron = { version = "^0.4.0", optional = true }
 serde = { version = "^0.8.21", optional = true }
