+++ src/integrations/mod.rs
@@ -1,2 +1,3 @@
 #[cfg(feature="iron-handlers")] pub mod iron_handlers;
+#[cfg(feature="rustc-serialize")] pub mod rustc_serialize;
 #[cfg(feature="serde")] pub mod serde;
